Sequence Types

BioSequences exports an abstract BioSequence type, and several concrete sequence types which inherit from it.

The abstract BioSequence

BioSequences provides an abstract type called a BioSequence{A<:Alphabet}. This abstract type, and the methods and traits is supports, allows for many algorithms in BioSequences to be written as generically as possible, thus reducing the amount of code to read and understand, whilst maintaining high performance when such code is compiled for a concrete BioSequence subtype. Additionally, it allows new types to be implemented that are fully compatible with the rest of BioSequences, providing that key methods or traits are defined).

This abstract type is parametric over concrete types of Alphabet, which define the range of symbols permitted in the sequence.

Some aliases are also provided for your convenience:

Type alias	Type

NucleotideSeq	BioSequence{<:NucleicAcidAlphabet}

AminoAcidSeq	BioSequence{AminoAcidAlphabet}

Any concrete sequence type compatible with BioSequences must inherit from BioSequence{A}, where A is the alphabet of the concrete sequence type. It must also have the following methods defined for it:

BioSequences.encoded_data — Function

Return the data member of seq that stores the encoded sequence data.

Base.length — Method

Get the length of a biological sequence.

If these requirements are satisfied, the following key traits and methods backing the BioSequences interface, should be defined already for the sequence type.

BioSequences.encoded_data_type — Function

Get the vector of bits storing a sequences packed encoded elements.

BioSequences.encoded_data_eltype — Function

Get the element type of the vector of bits storing a sequences packed encoded elements.

BioSequences.Alphabet — Method

Return the Alpahbet type that defines the biological symbols allowed for seq.

BioSymbols.alphabet — Method

Gets the alphabet encoding of a given BioSequence.

BioSequences.BitsPerSymbol — Type

The number of bits required to represent a packed symbol in a vector of bits.

BioSequences.bits_per_symbol — Function

Get the number of bits each symbol packed into a BioSequence uses, as an integer value.

As a result, the vast majority of methods described in the rest of this manual should work out of the box for the concrete sequence type. But they can always be overloaded if needed.

Long Sequences

Many genomics scripts and tools benefit from an efficient general purpose sequence type that allows you to create and edit sequences. In BioSequences, the LongSequence type fills this requirement.

LongSequence{A<:Alphabet} <: BioSequence{A} is parameterized by a concrete Alphabet type A that defines the domain (or set) of biological symbols permitted. For example, AminoAcidAlphabet is associated with AminoAcid and hence an object of the LongSequence{AminoAcidAlphabet} type represents a sequence of amino acids. Symbols from multiple alphabets can't be intermixed in one sequence type.

The following table summarizes common LongSequence types that have been given aliases for convenience.

Type	Symbol type	Type alias

LongSequence{DNAAlphabet{4}}	DNA	LongDNASeq

LongSequence{RNAAlphabet{4}}	RNA	LongRNASeq

LongSequence{AminoAcidAlphabet}	AminoAcid	LongAminoAcidSeq

The LongDNASeq and LongRNASeq aliases use a DNAAlphabet{4}, which means the sequence may store ambiguous nucleotides. If you are sure that nucleotide sequences store unambiguous nucleotides only, you can reduce the memory required by sequences by using a slightly different parameter: DNAAlphabet{2} is an alphabet that uses two bits per base and limits to only unambiguous nucleotide symbols (ACGT in DNA and ACGU in RNA). Replacing LongSequence{DNAAlphabet{4}} in your code with LongSequence{DNAAlphabet{2}} is all you need to do in order to benefit. Some computations that use bitwise operations will also be dramatically faster.

Sequence views

Similar to how Base Julia offers views of array objects, BioSequences offers view of LongSequences - the LongSubSeq{A<:Alphabet}.

Conceptually, a LongSubSeq{A} is similar to a LongSequence{A}, but instead of storing their own data, they refer to the data of a LongSequence. Modiying the LongSequence will be reflected in the view, and vice versa. If the underlying LongSequence is truncated, the behaviour of a view is undefined. For the same reason, some operations are not supported for views, such as resizing.

The purpose of LongSubSeq is that, since they only contain a pointer to the underlying array, an offset and a length, they are much lighter than LongSequences, and will be stack allocated on Julia 1.5 and newer. Thus, the user may construct millions of views without major performace implications.

Alphabet types

BioSequences.Alphabet — Type

Alphabets of biological symbols.

Alphabet is perhaps the most important type trait for biological sequences in BioSequences.jl.

An Alphabet represents a domain of biological symbols.

For example, DNAAlphabet{2} has a domain of unambiguous nucleotides (i.e. A, C, G, and T).

Alphabet types restrict and define the set of biological symbols, that can be encoded in a given biological sequence type. They ALSO define HOW that encoding is done.

An Alphabet type defines the encoding of biological symbols with a pair of associated encoder and decoder methods. These paired methods map between biological symbol values and a binary representation of the symbol.

Any type A <: Alphabet, is expected to implement the Base.eltype method for itself. It is also expected to implement the BitsPerSymbol method.

Alphabets control how biological symbols are encoded and decoded. They also confer many of the automatic traits and methods that any subtype of T<:BioSequence{A<:Alphabet} will get.