Abstract Types

BioSequences exports an abstract BioSequence type, and several concrete sequence types which inherit from it.

The abstract BioSequence

BioSequences provides an abstract type called a BioSequence{A<:Alphabet}. This abstract type, and the methods and traits is supports, allows for many algorithms in BioSequences to be written as generically as possible, thus reducing the amount of code to read and understand, whilst maintaining high performance when such code is compiled for a concrete BioSequence subtype. Additionally, it allows new types to be implemented that are fully compatible with the rest of BioSequences, providing that key methods or traits are defined).

BioSequences.BioSequence — Type

BioSequence{A <: Alphabet}

BioSequence is the main abstract type of BioSequences. Its subtypes are characterized by:

Being a linear container type with random access and indices Base.OneTo(length(x)).

Containing zero or more internal data elements of type encoded_data_eltype(typeof(x)).

Being associated with an Alphabet, A by being a subtype of BioSequence{A}.

A BioSequence{A} is indexed by an integer. The biosequence subtype, the index and the alphabet A determine how to extract the internal encoded data. The alphabet decides how to decode the data to the element type of the biosequence. Hence, the element type and container type of a BioSequence are separated.

Subtypes T of BioSequence must implement the following, with E begin an encoded data type:

Base.length(::T)::Int

encoded_data_eltype(::Type{T})::Type{E}

extract_encoded_element(::T, ::Integer)::E

copy(::T)

T must be able to be constructed from any iterable with length defined and

with a known, compatible element type.

Furthermore, mutable sequences should implement

encoded_setindex!(::T, ::E, ::Integer)

T(undef, ::Int)

resize!(::T, ::Int)

For compatibility with existing Alphabets, the encoded data eltype must be UInt.

Some aliases for BioSequence are also provided for your convenience:

BioSequences.NucleotideSeq — Type

An alias for BioSequence{<:NucleicAcidAlphabet}

BioSequences.AminoAcidSeq — Type

An alias for BioSequence{AminoAcidAlphabet}

Let's have a closer look at some of those methods that a subtype of BioSequence must implement. Check out julia base library docs for length, copy and resize!.

BioSequences.encoded_data_eltype — Function

encoded_data_eltype(::Type{<:BioSequence})

Returns the element type of the encoded data of the BioSequence. This is the return type of extract_encoded_element, i.e. the data type that stores the biological symbols in the biosequence.

See also: BioSequence

BioSequences.extract_encoded_element — Function

extract_encoded_element(::BioSequence{A}, i::Integer)

Returns the encoded element at position i. This data can be decoded using decode(A(), data) to yield the element type of the biosequence.

See also: BioSequence

BioSequences.encoded_setindex! — Function

encoded_setindex!(seq::BioSequence, x::E, i::Integer)

Given encoded data x of type encoded_data_eltype(typeof(seq)), sets the internal sequence data at the given index.

See also: BioSequence

A correctly defined subtype of BioSequence that satisfies the interface, will find the vast majority of methods described in the rest of this manual should work out of the box for that type. But they can always be overloaded if needed. Indeed the LongSequence type overloads Indeed some of the generic BioSequence methods, are overloaded for LongSequence, for example for transformation and counting operations where efficiency gains can be made due to the specific internal representation of a specific type.

The abstract Alphabet

Alphabets control how biological symbols are encoded and decoded. They also confer many of the automatic traits and methods that any subtype of T<:BioSequence{A<:Alphabet} will get.

BioSequences.Alphabet — Type

Alphabets of biological symbols.

Alphabet is the most important type trait for BioSequence. An Alphabet represents a set of biological symbols encoded by a sequence, e.g. A, C, G and T for a DNA Alphabet that requires only 2-bits to represent each symbol.

Subtypes of Alphabet are singleton structs that may or may not be parameterized.

Alphabets span over a finite set of biological symbols.

The alphabet controls the encoding from some internal "encoded data" to a BioSymbol

of the alphabet's element type, as well as the decoding, the inverse process.

An Alphabet's encode and decode methods must not produce invalid data.

Every subtype A of Alphabet must implement:

Base.eltype(::Type{A})::Type{S} for some eltype S, which must be a BioSymbol.

symbols(::A)::Tuple{Vararg{S}}. This gives tuples of all symbols in the set of A.

encode(::A, ::S)::E encodes a symbol to an internal data eltype E.

decode(::A, ::E)::S decodes an internal data eltype E to a symbol S.

Except for eltype which must follow Base conventions, all functions operating

on Alphabet should operate on instances of the alphabet, not the type.

If you want interoperation with existing subtypes of BioSequence, the encoded representation E must be of type UInt, and you must also implement:

BitsPerSymbol(::A)::BitsPerSymbol{N}, where the N must be zero

or a power of two in [1, 2, 4, 8, 16, 32, [64 for 64-bit systems]].

Concrete types

Implemented alphabets

Long Sequences

Many genomics scripts and tools benefit from an efficient general purpose sequence type that allows you to create and edit sequences. In BioSequences, the LongSequence type fills this requirement.

LongSequence{A<:Alphabet} <: BioSequence{A} is parameterized by a concrete Alphabet type A that defines the domain (or set) of biological symbols permitted. For example, AminoAcidAlphabet is associated with AminoAcid and hence an object of the LongSequence{AminoAcidAlphabet} type represents a sequence of amino acids. Symbols from multiple alphabets can't be intermixed in one sequence type.

The following table summarizes common LongSequence types that have been given aliases for convenience.

Type	Symbol type	Type alias

LongSequence{DNAAlphabet{4}}	DNA	LongDNASeq

LongSequence{RNAAlphabet{4}}	RNA	LongRNASeq

LongSequence{AminoAcidAlphabet}	AminoAcid	LongAminoAcidSeq

The LongDNASeq and LongRNASeq aliases use a DNAAlphabet{4}, which means the sequence may store ambiguous nucleotides. If you are sure that nucleotide sequences store unambiguous nucleotides only, you can reduce the memory required by sequences by using a slightly different parameter: DNAAlphabet{2} is an alphabet that uses two bits per base and limits to only unambiguous nucleotide symbols (ACGT in DNA and ACGU in RNA). Replacing LongSequence{DNAAlphabet{4}} in your code with LongSequence{DNAAlphabet{2}} is all you need to do in order to benefit. Some computations that use bitwise operations will also be dramatically faster.

Sequence views

Similar to how Base Julia offers views of array objects, BioSequences offers view of LongSequences - the LongSubSeq{A<:Alphabet}.

Conceptually, a LongSubSeq{A} is similar to a LongSequence{A}, but instead of storing their own data, they refer to the data of a LongSequence. Modiying the LongSequence will be reflected in the view, and vice versa. If the underlying LongSequence is truncated, the behaviour of a view is undefined. For the same reason, some operations are not supported for views, such as resizing.

The purpose of LongSubSeq is that, since they only contain a pointer to the underlying array, an offset and a length, they are much lighter than LongSequences, and will be stack allocated on Julia 1.5 and newer. Thus, the user may construct millions of views without major performance implications.